+++ CHANGELOG.md
@@ -1,4 +1,23 @@
 # Change Log
+<!--
+## [1.0.1]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+### Updated 
+-->
+
+<!-- ## [v-inc] ${YEAR4}-${MONTHNUMBER}-${DATE} -->
+
+## [0.0.2] 2022-04-05
+### Added
+- Illegal Opcodes
+### Updated 
+- HTML Content
+- Readme.md
 
 All notable changes to the "c64opcodes" extension will be documented in this file.